Refactor algorithm enums
